Move `InvertibleAdaptersMixin` to separate module (fixes docs building issue)
