+++ src/transformers/adapters/models/bert.py
@@ -15,8 +15,9 @@ from ..heads import (
     QuestionAnsweringHead,
     TaggingHead,
 )
+from ..invertible_adapters_mixin import InvertibleAdaptersMixin
 from ..layer import AdapterLayerBaseMixin
-from ..model_mixin import InvertibleAdaptersMixin, ModelAdaptersMixin
+from ..model_mixin import ModelAdaptersMixin
 
 
 logger = logging.getLogger(__name__)
