+++ src/transformers/adapters/models/distilbert.py
@@ -4,7 +4,8 @@ import torch
 from torch import nn
 
 from ..composition import AdapterCompositionBlock, parse_composition
-from ..model_mixin import InvertibleAdaptersMixin, ModelAdaptersMixin
+from ..invertible_adapters_mixin import InvertibleAdaptersMixin
+from ..model_mixin import ModelAdaptersMixin
 from .bert import BertEncoderAdaptersMixin, BertModelHeadsMixin, BertOutputAdaptersMixin, BertSelfOutputAdaptersMixin
