+++ src/transformers/models/bart/modeling_bart.py
@@ -25,7 +25,8 @@ from torch import nn
 from torch.nn import CrossEntropyLoss, MSELoss
 
 from ...activations import ACT2FN
-from ...adapters.model_mixin import InvertibleAdaptersMixin, ModelWithHeadsAdaptersMixin
+from ...adapters.invertible_adapters_mixin import InvertibleAdaptersMixin
+from ...adapters.model_mixin import ModelWithHeadsAdaptersMixin
 from ...adapters.models.bart import (
     BartDecoderLayerAdaptersMixin,
     BartEncoderDecoderAdaptersMixin,
